weather_and_time moved Time to its own file weather class now attached to game_map instead of Engine annotations for components
