+++ src/components/sails.py
@@ -1,6 +1,12 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from components.base import BaseComponent
 from constants.colors import colors
-from entity import Actor
+
+if TYPE_CHECKING:
+    from entity import Actor
 
 
 class Sails(BaseComponent):
